[BugFix] Fix envpool (#1530)
